+++ app/api/district/model.ts
@@ -1,0 +1,31 @@
+import { DataTypes } from 'sequelize'
+import db from '../../../config/db'
+// Define schema
+export const districtSchema = db.define(
+  'district',
+  {
+    // Model attributes are defined here
+    id: {
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    city_code: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    code: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  },
+  {
+    // Other model options go here
+    tableName: 'indonesia_districts',
+    timestamps: false,
+  },
+)